Reconfigure interactive CLI into core plugin
