+++ lib/plugins/Plugins.json
@@ -1,5 +1,6 @@
 {
   "plugins": [
+    "./interactiveCli",
     "./config/config.js",
     "./create/create.js",
     "./install/install.js",
