+++ lib/plugins/interactiveCli/utils.js
@@ -1,0 +1,14 @@
+'use strict';
+
+const inquirer = require('inquirer');
+
+module.exports = {
+  confirm: message =>
+    inquirer
+      .prompt({
+        message,
+        type: 'confirm',
+        name: 'isConfirmed',
+      })
+      .then(({ isConfirmed }) => isConfirmed),
+};
